Refactort protocol
