--- pkg/stream/protocol.go
@@ -1,17 +1,0 @@
-package stream
-
-type StreamingResponse struct {
-	FrameLen          uint32
-	CommandID         uint16
-	Key               uint16
-	Version           uint16
-	CorrelationId     uint32
-	ResponseCode      uint16
-	PublishID         uint8
-	PublishingIdCount uint64
-}
-
-type WriteResponse struct {
-	Code int
-	Err  error
-}
